+++ src/bpftrace.h
@@ -187,6 +187,7 @@ public:
   const FuncsModulesMap &get_traceable_funcs() const;
   KConfig kconfig;
   std::vector<std::unique_ptr<AttachedProbe>> attached_probes_;
+  std::unique_ptr<AttachedProbe> sigusr1_attached_probe_;
 
   std::map<std::string, std::unique_ptr<PCAPwriter>> pcap_writers;
 
@@ -229,9 +230,10 @@ public:
   Config config_;
 
 private:
-  int run_special_probe(std::string name,
-                        BpfBytecode &bytecode,
-                        void (*trigger)(void));
+  std::optional<std::unique_ptr<AttachedProbe>> attach_special_probe(
+      std::string name,
+      BpfBytecode &bytecode);
+  int run_special_probe(AttachedProbe &ap, void (*trigger)(void));
   void *ksyms_{ nullptr };
   // note: exe_sym_ is used when layout is same for all instances of program
   std::map<std::string, std::pair<int, void *>> exe_sym_; // exe -> (pid, cache)